moved Loader componenet to loaders folder
